pyproject.toml
[tool.poetry]
name = "extract-facebook-ads"
version = "0.1.0"
description = ""
authors = ["jonhnatta <jonhnatta.santos@gmail.comclear>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.32.3"
pyspark = "^3.5.5"
python-dotenv = "^1.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
